+++ CHANGELOG.md
@@ -14,13 +14,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### C# -> VB
 
 
-## [8.2.0] - 2020-10-12
+## [8.2.1] - 2020-10-28
 
-
-* Web UI improvements [#644](https://github.com/icsharpcode/CodeConverter/pull/644)
 
 ### Vsix
 
+
+### VB -> C#
+* XML output path needs output path prepending [#641](https://github.com/icsharpcode/CodeConverter/issues/641)
+* Fix issue when converting `?.` operator - [#673](https://github.com/icsharpcode/CodeConverter/issues/673)
+
+### C# -> VB
+* XML output path needs output path prepending [#641](https://github.com/icsharpcode/CodeConverter/issues/641)
+* Converting explicit and implicit operator failed [#659](https://github.com/icsharpcode/CodeConverter/issues/659)
+
+## [8.2.0] - 2020-10-12
+* Web UI improvements [#644](https://github.com/icsharpcode/CodeConverter/pull/644)
+
+### Vsix
 * Add "Paste as VB/C#" to "Paste special" menu [#622](https://github.com/icsharpcode/CodeConverter/pull/622)